pyproject.toml
[tool.poetry]
name = "example-ed-tech"
version = "0.1.0"
description = ""
authors = ["Dzhurunov Tamerlan <dzhurunovt@gmail.com>"]
readme = "README.md"
packages = [{include = "example_ed_tech"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
psycopg2 = "^2.9.5"
python-environ = "^0.4.54"
djangorestframework = "^3.14.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
